Update search functionality and move converters
