+++ FluentFlyoutWPF/Classes/Converters/BoolToFullHalfOpacityConverter.cs
@@ -2,15 +2,14 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 using System.Globalization;
-using System.Windows;
 using System.Windows.Data;
 
-namespace FluentFlyoutWPF.Classes.Utils;
+namespace FluentFlyoutWPF.Classes.Converters;
 
-public class BoolToVisibleCollapsedConverter : IValueConverter
+public class BoolToFullHalfOpacityConverter : IValueConverter
 {
-    public Visibility TrueValue { get; set; } = Visibility.Visible;
-    public Visibility FalseValue { get; set; } = Visibility.Collapsed;
+    public double TrueValue { get; set; } = 1;
+    public double FalseValue { get; set; } = 0.5;
 
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
@@ -27,9 +26,9 @@ public class BoolToVisibleCollapsedConverter : IValueConverter
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
     {
-        if (value is Visibility visibility)
+        if (value is double opacity)
         {
-            return visibility == TrueValue;
+            return opacity == TrueValue;
         }
         return false;
     }
