+++ FluentFlyoutWPF/Classes/Converters/BoolToVisibleCollapsedConverter.cs
@@ -2,14 +2,15 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 using System.Globalization;
+using System.Windows;
 using System.Windows.Data;
 
-namespace FluentFlyoutWPF.Classes.Utils;
+namespace FluentFlyoutWPF.Classes.Converters;
 
-public class BoolToFullHalfOpacityConverter : IValueConverter
+public class BoolToVisibleCollapsedConverter : IValueConverter
 {
-    public double TrueValue { get; set; } = 1;
-    public double FalseValue { get; set; } = 0.5;
+    public Visibility TrueValue { get; set; } = Visibility.Visible;
+    public Visibility FalseValue { get; set; } = Visibility.Collapsed;
 
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
@@ -26,9 +27,9 @@ public class BoolToFullHalfOpacityConverter : IValueConverter
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
     {
-        if (value is double opacity)
+        if (value is Visibility visibility)
         {
-            return opacity == TrueValue;
+            return visibility == TrueValue;
         }
         return false;
     }
