+++ FluentFlyoutWPF/Classes/Converters/DecimalPercentageToFullConverter.cs
@@ -4,7 +4,7 @@
 using System.Globalization;
 using System.Windows.Data;
 
-namespace FluentFlyoutWPF.Classes.Utils;
+namespace FluentFlyoutWPF.Classes.Converters;
 
 public class DecimalPercentageToFullConverter : IValueConverter
 {
