+++ FluentFlyoutWPF/Pages/VolumeMixerPage.xaml
@@ -6,7 +6,7 @@
       xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
       xmlns:controls="clr-namespace:MicaWPF.Controls;assembly=MicaWPF"
       xmlns:viewModels="clr-namespace:FluentFlyoutWPF.ViewModels"
-      xmlns:utils="clr-namespace:FluentFlyoutWPF.Classes.Utils"
+      xmlns:converters="clr-namespace:FluentFlyoutWPF.Classes.Converters"
       xmlns:fluentflyoutcontrols="clr-namespace:FluentFlyout.Controls"
       mc:Ignorable="d"
       d:DesignHeight="800" d:DesignWidth="800"
@@ -16,8 +16,8 @@
       d:Background="White">
 
     <Page.Resources>
-        <utils:BoolToVisibleCollapsedConverter x:Key="BoolToCollapsedVisibleConverter" TrueValue="Collapsed" FalseValue="Visible" />
-        <utils:BoolToFullHalfOpacityConverter x:Key="BoolToFullHalfOpacityConverter" />
+        <converters:BoolToVisibleCollapsedConverter x:Key="BoolToCollapsedVisibleConverter" TrueValue="Collapsed" FalseValue="Visible" />
+        <converters:BoolToFullHalfOpacityConverter x:Key="BoolToFullHalfOpacityConverter" />
     </Page.Resources>
 
     <StackPanel MaxWidth="1000" Margin="16">
